refactor(examples/web-linearlite): extract KeyboardShortcuts to separate file
Move KeyboardShortcuts component from __root.tsx to its own file in
components/common for better code organization.
