+++ examples/web-linearlite/src/components/common/keyboard-shortcuts.tsx
@@ -1,0 +1,32 @@
+import { useRouter } from '@tanstack/react-router'
+import React from 'react'
+import { useAppStore } from '../../livestore/store.ts'
+import { NewIssueModalContext } from '../../app/contexts.ts'
+
+export const KeyboardShortcuts = () => {
+  const store = useAppStore()
+  const router = useRouter()
+  const { setNewIssueModalStatus } = React.useContext(NewIssueModalContext)!
+
+  React.useEffect(() => {
+    const handleKeyDown = (e: KeyboardEvent) => {
+      const element = e.target as HTMLElement
+      if (element.classList.contains('input')) return
+
+      if (e.key === 'c') {
+        setNewIssueModalStatus(0)
+        e.preventDefault()
+      }
+
+      // We use '?' instead of '/' because Shift+/ is '?'
+      if (e.key === '?' && e.shiftKey) {
+        router.navigate({ to: '/$storeId/search', params: { storeId: store.storeId } })
+        e.preventDefault()
+      }
+    }
+    window.addEventListener('keydown', handleKeyDown)
+    return () => window.removeEventListener('keydown', handleKeyDown)
+  }, [router, store.storeId, setNewIssueModalStatus])
+
+  return null
+}
